--- a/app/org/maproulette/framework/controller/TaskController.scala
+++ b/app/org/maproulette/framework/controller/TaskController.scala
@@ -77,10 +77,10 @@ class TaskController @Inject() (
   /**
     * Gets all the tasks within a bounding box
     *
-    * @param left   The minimum latitude for the bounding box
-    * @param bottom The minimum longitude for the bounding box
-    * @param right  The maximum latitude for the bounding box
-    * @param top    The maximum longitude for the bounding box
+    * @param left   The minimum longitude for the bounding box
+    * @param bottom The minimum latitude for the bounding box
+    * @param right  The maximum longitude for the bounding box
+    * @param top    The maximum latitude for the bounding box
     * @param limit  Limit for the number of returned tasks
     * @param offset The offset used for paging
     * @return
@@ -125,10 +125,10 @@ class TaskController @Inject() (
   /**
     * Gets all the task markers within a bounding box
     *
-    * @param left   The minimum latitude for the bounding box
-    * @param bottom The minimum longitude for the bounding box
-    * @param right  The maximum latitude for the bounding box
-    * @param top    The maximum longitude for the bounding box
+    * @param left   The minimum longitude for the bounding box
+    * @param bottom The minimum latitude for the bounding box
+    * @param right  The maximum longitude for the bounding box
+    * @param top    The maximum latitude for the bounding box
     * @param limit  Limit for the number of returned tasks
     * @return
     */