… app/org/maproulette/models/dal/ChallengeDAL.scala
… conf/v2_route/task.api