The ` PUT     /tasks/box/:left/:bottom/:right/:top ` endpoint has some unneeded complexity when fetching for task markers. The reason why that endpoint was so complex is because it is also used to fetch task table data. This led to slower performance when fetching markers because paging, ordering, and sorting was applied.

Because that structure is needed for the tables but not for fetching tasks markers, I have created a new endpoint to reduce complexity: `PUT     /tasks/markers/:left/:bottom/:right/:top `.

This new endpoint exports the same task data as the previous endpoint, but doesn't query for the total task count, and also doesn't do any ordering or paging.

21771 task in view with 1001 task limit:
Old endpoint (Time to complete: ~ 1.1s-1.9s ):
<img width="1453" alt="Screenshot 2024-07-19 at 4 36 15 PM" src="https://github.com/user-attachments/assets/9ec9f0d2-25a5-4e6e-a14a-f2b7631616d6">

New endpoint  (Time to complete: ~ 400ms-600ms):
<img width="1453" alt="Screenshot 2024-07-19 at 4 36 33 PM" src="https://github.com/user-attachments/assets/df1354b6-ee05-44fd-bfae-e5f92f5a3eb8">

977 task in view with 1001 task limit:
Old endpoint  (Time to complete: ~ 450ms-700ms):
<img width="1455" alt="Screenshot 2024-07-19 at 4 37 34 PM" src="https://github.com/user-attachments/assets/42dd79db-e536-4afd-a346-3ae69e3bfe95">

New endpoint  (Time to complete: ~300ms-550ms):
<img width="1454" alt="Screenshot 2024-07-19 at 4 37 59 PM" src="https://github.com/user-attachments/assets/10ae372d-c40c-4b32-9579-e7f3002e3ef1">


Times were recorded on the same device at a similar time, and ran multiple times to get an average. There is a small performance increase with the new endpoint when the number of tasks in view is below the limit. A bigger performance increase is noticeable whenever there are significantly more than 1001 tasks(or the limit) as the endpoint no longer needs to query for all the tasks the meet the filter requirements to get the total tssk count.